pyproject.toml
[project]
name = "steam-translator"
version = "0.1.0"
description = "Steam游戏汉化工具 - 实时翻译覆盖 + 汉化补丁制作"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "PySide6>=6.7",
    "mss>=9.0",
    "httpx>=0.27",
    "Pillow>=10.0",
    "numpy>=1.26",
    "pynput>=1.7",
    "cachetools>=5.3",
]

[project.optional-dependencies]
ocr = [
    "rapidocr-onnxruntime>=1.3",
]
ocr-paddle = [
    "paddlepaddle>=2.6",
    "paddleocr>=2.7",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
steam-translator = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]
